refactor distance handling
